requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
h11==0.14.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2

uvicorn==0.32.0
fastapi==0.115.4
python-jose
sqlalchemy
PyJWT
pydantic[email]
requests
python-multipart
google-cloud-speech
OpenAI